LayoutTests/http/tests/cookies/only-accept-first-party-cookies-expected.txt

Tests the first-party-only cookie policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame2-->'
--------
Cookies are:

--------
Frame: '<!--frame3-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame4-->'
--------
